digit-bot-main/requirements.txt
numpy
opencv-python
scikit-learn
scikit-image
torch
torchvision
Pillow
joblib
matplotlib